pyproject.toml
[project]
name = "ibf"
version = "0.3.0"
description = "Unified impact-based forecast CLI and service"
readme = "README.md"
authors = [
    { name = "tehoro", email = "tehoro@users.noreply.github.com" }
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "click>=8.1.7,<8.2",
    "typer>=0.12.5,<0.13",
    "rich>=13.9.4,<14",
    "pydantic>=2.9.2,<3.0.0",
    "python-dotenv>=1.0.1,<2.0",
    "requests>=2.32.3,<3.0",
    "beautifulsoup4>=4.12.3,<5.0",
    "feedparser>=6.0.11,<7.0",
    "openai>=1.53.0,<2.0",
    "google-generativeai>=0.8.4,<0.9",
    "arrow>=1.3.0,<2.0",
    "pytz>=2023.4",
    "shapely>=2.0.6,<3.0",
    "lxml>=5.3.0,<6.0",
    "numpy>=1.26.4,<3.0",
    "h3>=4.1.0,<5.0",
    "folium>=0.16.0,<0.17",
    "selenium>=4.24.0,<5.0",
    "staticmap>=0.5.5,<0.6",
    "Pillow>=10.4.0,<11.0",
    "timezonefinder>=6.4.0,<7.0",
    "pytest>=8.3.3,<9.0",
    "pytest-mock>=3.14.0,<4.0",
    "google-genai>=1.56.0",
]

[project.optional-dependencies]
dev = []

[project.scripts]
ibf = "ibf.cli:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
